Warm start refactoring
- Go upgraded to 1.16
- Embedded scripts moved to separate files using go:embed
- Refactored init scripts to be reused for warm start
- Removed peer finder - peers list computed by the operator so that warm
start can also use it - as a side effect much faster pod start
- Added back init container docker files
- Refactored config map code to be in a single file
