+++ controllers/configmap.go
@@ -1,0 +1,190 @@
+package controllers
+
+import (
+	"bytes"
+	"embed"
+	"encoding/json"
+	"fmt"
+	"io/fs"
+	"path/filepath"
+	"strings"
+	"text/template"
+
+	asdbv1alpha1 "github.com/aerospike/aerospike-kubernetes-operator/api/v1alpha1"
+	"github.com/aerospike/aerospike-kubernetes-operator/pkg/utils"
+	"github.com/aerospike/aerospike-management-lib/asconfig"
+	ctrl "sigs.k8s.io/controller-runtime"
+)
+
+var pkglog = ctrl.Log.WithName("lib.asconfig")
+
+const (
+	// AerospikeTemplateConfFileName is the name of the aerospike conf template
+	AerospikeTemplateConfFileName = "aerospike.template.conf"
+
+	// NetworkPolicyHashFileName stores the network policy hash
+	NetworkPolicyHashFileName = "networkPolicyHash"
+
+	// PodSpecHashFileName stores the pod spec hash
+	PodSpecHashFileName = "podSpecHash"
+
+	// AerospikeConfHashFileName stores the Aerospike config hash
+	AerospikeConfHashFileName = "aerospikeConfHash"
+)
+
+type initializeTemplateInput struct {
+	WorkDir         string
+	MultiPodPerHost bool
+	NetworkPolicy   asdbv1alpha1.AerospikeNetworkPolicy
+	PodPort         int32
+	PodTLSPort      int32
+}
+
+//go:embed scripts
+var scripts embed.FS
+
+// A map from script name to its template.
+var scriptTemplates = make(map[string]*template.Template)
+
+func init() {
+	err := fs.WalkDir(scripts, ".", func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if !d.IsDir() {
+			content, err := fs.ReadFile(scripts, path)
+			if err != nil {
+				return err
+			}
+
+			evaluated, err := template.New(path).Parse(string(content))
+			if err != nil {
+				return err
+			}
+			key := filepath.Base(path)
+			scriptTemplates[key] = evaluated
+		}
+		return nil
+	})
+
+	if err != nil {
+		// Error reading embedded script templates.
+		panic(fmt.Sprintf("error reading embedded script templates: %v", err))
+	}
+}
+
+// CreateConfigMapData create configMap data
+func (r *AerospikeClusterReconciler) CreateConfigMapData(aeroCluster *asdbv1alpha1.AerospikeCluster, rack asdbv1alpha1.Rack) (map[string]string, error) {
+	// Add config template
+	confTemp, err := r.buildConfigTemplate(aeroCluster, rack)
+	if err != nil {
+		return nil, fmt.Errorf("failed to build config template: %v", err)
+	}
+
+	// Add conf file
+	confData, err := r.getBaseConfData(aeroCluster, rack)
+	if err != nil {
+		return nil, fmt.Errorf("failed to build config template: %v", err)
+	}
+	confData[AerospikeTemplateConfFileName] = confTemp
+
+	// Add conf hash
+	confHash, err := utils.GetHash(confTemp)
+	if err != nil {
+		return nil, err
+	}
+	confData[AerospikeConfHashFileName] = confHash
+
+	// Add networkPolicy hash
+	policy := aeroCluster.Spec.AerospikeNetworkPolicy
+	policyStr, err := json.Marshal(policy)
+	if err != nil {
+		return nil, err
+	}
+	policyHash, err := utils.GetHash(string(policyStr))
+	if err != nil {
+		return nil, err
+	}
+	confData[NetworkPolicyHashFileName] = policyHash
+
+	// Add podSpec hash
+	podSpec := aeroCluster.Spec.PodSpec
+	podSpecStr, err := json.Marshal(podSpec)
+	if err != nil {
+		return nil, err
+	}
+	podSpecHash, err := utils.GetHash(string(podSpecStr))
+	if err != nil {
+		return nil, err
+	}
+	confData[PodSpecHashFileName] = podSpecHash
+
+	return confData, nil
+}
+
+func (r *AerospikeClusterReconciler) buildConfigTemplate(aeroCluster *asdbv1alpha1.AerospikeCluster, rack asdbv1alpha1.Rack) (string, error) {
+	log := pkglog.WithValues("aerospikecluster", utils.ClusterNamespacedName(aeroCluster))
+
+	version := strings.Split(aeroCluster.Spec.Image, ":")
+
+	configMap := rack.AerospikeConfig.Value
+	log.V(1).Info("AerospikeConfig", "config", configMap, "image", aeroCluster.Spec.Image)
+
+	asConf, err := asconfig.NewMapAsConfig(version[1], configMap)
+	if err != nil {
+		return "", fmt.Errorf("failed to load config map by lib: %v", err)
+	}
+
+	// No need for asConf version validation, it's already validated in admission webhook
+
+	confFile := asConf.ToConfFile()
+	log.V(1).Info("AerospikeConfig", "conf", confFile)
+
+	return confFile, nil
+}
+
+// getBaseConfData returns the basic data to be used in the config map for input aeroCluster spec.
+func (r *AerospikeClusterReconciler) getBaseConfData(aeroCluster *asdbv1alpha1.AerospikeCluster, rack asdbv1alpha1.Rack) (map[string]string, error) {
+
+	workDir := asdbv1alpha1.GetWorkDirectory(rack.AerospikeConfig)
+
+	// Include initialization and restart scripts
+	initializeTemplateInput := initializeTemplateInput{
+		WorkDir:         workDir,
+		MultiPodPerHost: aeroCluster.Spec.MultiPodPerHost,
+		NetworkPolicy:   aeroCluster.Spec.AerospikeNetworkPolicy,
+		PodPort:         asdbv1alpha1.ServicePort,
+		PodTLSPort:      asdbv1alpha1.ServiceTLSPort,
+	}
+
+	baseConfData := map[string]string{}
+	for path, template := range scriptTemplates {
+		var script bytes.Buffer
+		err := template.Execute(&script, initializeTemplateInput)
+		if err != nil {
+			return nil, err
+		}
+
+		baseConfData[path] = script.String()
+	}
+
+	// Include peer list.
+	peers := getFQDNsForCluster(aeroCluster)
+
+	baseConfData["peers"] = strings.Join(peers, "\n")
+	return baseConfData, nil
+}
+
+func getFQDNsForCluster(aeroCluster *asdbv1alpha1.AerospikeCluster) []string {
+	podNames := []string{}
+	rackStateList := getNewRackStateList(aeroCluster)
+	for _, rackState := range rackStateList {
+		size := rackState.Size
+		stsName := getNamespacedNameForSTS(aeroCluster, rackState.Rack.ID)
+		for i := 0; i < size; i++ {
+			podNames = append(podNames, getFQDNForPod(aeroCluster,
+				getSTSPodName(stsName.Name, int32(i))))
+		}
+	}
+	return podNames
+}
